improve public API documentation
