+++ src/util/emitter.ts
@@ -3,6 +3,9 @@ export type Listener<Args extends unknown[] = unknown[]> = (
 ) => unknown;
 export type UnknownEvents = Record<string, unknown[]>;
 
+/**
+ * A class used to subscribe to and emit events
+ */
 export class Emitter<Events extends UnknownEvents = UnknownEvents> {
 	private collectable: Partial<{
 		[K in keyof Events]: Events[K][];
