+++ src/util/string-prefixes.ts
@@ -1,5 +1,11 @@
 import { StringRecordId, Uuid } from "../data";
 
+/**
+ * A template literal tag function for parsing a string type
+ * @param string - The string to parse
+ * @param values - The interpolated values
+ * @returns The parsed string
+ */
 export function s(
 	string: string[] | TemplateStringsArray,
 	...values: unknown[]
@@ -10,6 +16,12 @@ export function s(
 	);
 }
 
+/**
+ * A template literal tag function for parsing a string into a Date
+ * @param string - The string to parse
+ * @param values - The interpolated values
+ * @returns The parsed Date
+ */
 export function d(
 	string: string[] | TemplateStringsArray,
 	...values: unknown[]
@@ -17,6 +29,12 @@ export function d(
 	return new Date(s(string, values));
 }
 
+/**
+ * A template literal tag function for parsing a string into a StringRecordId
+ * @param string - The string to parse
+ * @param values - The interpolated values
+ * @returns The parsed StringRecordId
+ */
 export function r(
 	string: string[] | TemplateStringsArray,
 	...values: unknown[]
@@ -24,6 +42,12 @@ export function r(
 	return new StringRecordId(s(string, values));
 }
 
+/**
+ * A template literal tag function for parsing a string into a Uuid
+ * @param string - The string to parse
+ * @param values - The interpolated values
+ * @returns The parsed Uuid
+ */
 export function u(
 	string: string[] | TemplateStringsArray,
 	...values: unknown[]
